rename root.ts
