+++ app/src/api.ts
@@ -3,7 +3,7 @@ import {
   fetchBaseQuery,
   FetchBaseQueryError,
 } from "@reduxjs/toolkit/dist/query/react";
-import { RootState } from "./store/store";
+import { RootState } from "./store/root";
 import { DateTime } from "luxon";
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import {
