+++ app/src/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import App from "./App";
-import { store } from "./store/store";
+import { store } from "./store/root";
 
 ReactDOM.render(
   <React.StrictMode>
